rename translations
