+++ vllm/entrypoints/openai/speech_to_text/__init__.py
