+++ vllm/entrypoints/openai/speech_to_text/protocol.py
